Show a dialog rather than always downloading updates.
Summary:
Right now we always automatically download the latest version of Simple Radio
if we're compiled in dev mode and there's a new version available on the
buildbot server. Let's prompt the user and ask if they want to update instead.

Test Plan: Opened in the emulator, got a dialog, tested both buttons.

Reviewers: jcarter

Reviewed By: jcarter

Maniphest Tasks: T38

Differential Revision: https://phabricator.edgesysdesign.com/D11