Introduce a package object and stick our implicits in it.
Summary:
I am trying to clean up our `MainActivity` a bit, and this is a first step in
that direction.

This means that our implicits are global, but prevents having to explicitly
include them in every activity that we want to thread off (so we don't block the
UI thread), which I suspect will happen often.

Test Plan: Compiled and ran in the emulator with no changes from before.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D12