Turn PL tone select into a Spinner.
Summary:
Make PL tones a spinner, since they should never change.
We'll have to handle the `onItemSelected` eventually, which will set something
to store this.

This doesn't actually do anything right now.

Particularly, I think we should store radio state in an object later on, and
this can just change a var within that object (ugh mutability, but this might
be our best choice here). This also gives us the ability to serialize the result
and store it in a file so that we can restore last saved state when the app is
started.

Test Plan: Played with UI in the emulator.

Reviewers: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D14