The start of a Frequency class.
Summary:
This is (will be, eventually) a class for manipulating frequencies.
Right now it takes a string, parses it into a structure (List[Long]), and
provides a field, `.toHz` for getting it back as a `Long`.

Test Plan: Tested in REPL, but really this should be added to the test suite.

Reviewers: jcarter

Reviewed By: jcarter

Maniphest Tasks: T42

Differential Revision: https://phabricator.edgesysdesign.com/D15