Target Android 2.2 for now.
Summary:
We don't lose anything by targetting 2.2 for now, and until it causes a real
issue, I can't think of a reason not to. I don't think it's wise to "officially"
support it -- but then again our hardware will be running 4 and there's no need
to. But for now, in the early stages of development, if people running 2.2 want
to test our app, why not?

Test Plan: Very quickly tested in the emulator.

Reviewers: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D18