Move the Frequency library out of Simple Radio.
Summary:
This commit moves the Frequency library out of the Simple Radio repo and
adds that repository as a submodule here.

The advantage of doing a submodule as opposed to having SBT depend on the
git repository is that it's easier to update the submodule from a working
copy.

In fact, this commit includes an updater script which will pull the latest
Simple Radio + all submodules for you.

Test Plan: This will be easier to test after D20 lands.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D21