Show commit message and bold it, in the update dialog.
Summary:
This makes it easier for someone to decide whether or not to update.

This also requires a minor change to the buildsys script, so that the
commit-sha.txt is generated via `git log --oneline -1 --color=never`

This also introduces a workaround to
[[https://issues.scala-lang.org/browse/SI-5397 | this bug]] which might also
fix the issue in T39. Will test.

Test Plan: Tried in emulator.

Reviewers: jcarter

Reviewed By: jcarter

Maniphest Tasks: T39

Differential Revision: https://phabricator.edgesysdesign.com/D24