Add an empty class to fix a crash.
Summary:
When a widget is added or the time elapses and an update is triggered,
it tries to spawn an instance of the widget's receiver, which did not exist.

This fixes a crash on widget update and widget creation.

Test Plan: Created the widget in the emulator and didn't get a crash.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D25