Use the Frequency library for the Frequency field.
Summary:
This is more just a test to show that it's possible, but it's our first step
toward T54, in that we can now call `+()` and `-()` on the Frequency object
to change it with swipes.

This depends on D38.

Test Plan: Tried it in the emulator -- no behavior is currently different.

Reviewers: jcarter

Reviewed By: jcarter

Maniphest Tasks: T54

Differential Revision: https://phabricator.edgesysdesign.com/D39