Add a double tap listener to frequency field.
Summary:
This isn't a fling method, but it's a step in that direction, proving that it
is possible, and easy, to add a `GestureDetector` to the `EditText.

Right now, this crashes if you double tap when the frequency is over 1000 MHz
due some `Frequency` clownage which will be fixed when D40 lands.

Test Plan:
Played around in the emulator.

- It worked when expected and crashed when expected.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D41