Rework the auto update check a bit.
Summary:
On some cell providers, if your service is cut off (e.g. due to nonpayment (ask
me how I know that one)), they proxy all web requests off to their internal
portal page.

Simple Radio's updater didn't handle that, and ended up spewing a bunch of HTML
in the update request dialog, which is obviously horrible.

This diff fixes that by checking for a valid content type (text/plain). It also
cleans up some of the logic for dealing with the AlertDialog used to show this
information.

Test Plan:
Temporarily commented out the "is not an emulator" check, and tested in the
emulator.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D63