Add a "tap" implicit method.
Summary:
This method does the same thing that the Ruby method of the same name does.
It works around an annoyance with some APIs (including some Android APIs) whose
setters don't return the instance of the object they are controlling.

For example, Android's `EditText`:

  lang=scala
  val et = new EditText(context).setHint("some placeholder")

You'd except that to be an `EditText` instance with a hint set. But instead, it's
magically `Unit` (Scala's "Useless result" return value, akin to Java's `void`).

Instead, they expect you to do this:

  lang=scala
  val et = new EditText(context)
  et.setHint("some placeholder")
  et.setSingleLine(true)

...which works fine, but is unnecessarily verbose.

This diff adds a `tap` method, to work around this a bit.

  lang=scala
  import scala.language.reflectiveCalls
  val et = new EditText(context).tap { obj =>
    obj.setHint("some placeholder")
    obj.setSingleLine(true)
  }

This will cause `et` to have what we expect -- the `EditText`.

Test Plan: Tested this locally by changing "labelInput" from D65 to use `tap`.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D66