Move IQEncode into the SR repo for now.
Summary:
This program will take a raw audio file, and output another one, in stereo, using
I and Q samples.

It has a lot of room for improvement, and currently spawns two `ByteBuffer`
instances per sample, and we can optimize this out if we need to.

It also has a **lot** of mutability, which I would like to get rid of before we
actually add this to the SR app.

Test Plan:
Spent many hours working on this with @jcarter last night and got it 'working'.
After making some changes today, I ensured the result was the same as what was
considered "working" last night.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D64