Begin using Scaloid in `AboutActivity`.
Summary:
See some discussion in T73. Scaloid is a really nice looking library that
wraps a lot of the Android SDK in Scala convention.

It has a decent size following (233 stars on GitHub as of this writing), and has
come a **long** way in the month since I filed T73.

We can port to it slowly, without breaking the old code, which is what this diff
starts doing - it ports `AboutActivity` to use Scaloid instead of the Android
SBT stuff that the rest of SR uses.

Test Plan:
Compiled, ran in emulator, went to about view, double-tapped in the
easter egg place, and watch fun things happen.

Reviewers: jcarter

Reviewed By: jcarter

Maniphest Tasks: T73

Differential Revision: https://phabricator.edgesysdesign.com/D113